backend/app/db.py
import os
from neo4j import GraphDatabase
from dotenv import load_dotenv
load_dotenv()


_driver = None

def get_driver():
    global _driver
    if _driver is None:
        _driver = GraphDatabase.driver(
            os.environ["NEO4J_URI"],
            auth=(os.environ["NEO4J_USER"], os.environ["NEO4J_PASSWORD"]),
        )
    return _driver
