backend/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
neo4j==5.24.0
pandas==2.2.2
python-dotenv==1.0.1
httpx==0.27.2

pytest==8.3.2
pytest-cov==5.0.0
testcontainers==4.8.2
requests==2.32.3
